Moved createFormData to utils.
